@@ -15,7 +15,7 @@
   <description>Demo project for Spring Boot</description>
 
   <properties>
-    <java.version>11</java.version>
+    <java.version>12</java.version>
   </properties>
 
   <dependencies>
